pipelines/yc-scraper/requirements.txt
# HELM Pilot — YC Intelligence Pipeline
# Python 3.10+
psycopg2-binary>=2.9.10
